Dodanie klasy Product
