+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Product.java
@@ -1,0 +1,36 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.math.BigDecimal;
+import java.util.Date;
+
+public class Product {
+	private String productId;
+
+	private BigDecimal productPrice;
+
+	private String productName;
+
+	private Date productSnapshotDate;
+
+	private String productType;
+
+	public String getProductId() {
+		return productId;
+	}
+
+	public BigDecimal getProductPrice() {
+		return productPrice;
+	}
+
+	public String getProductName() {
+		return productName;
+	}
+
+	public Date getProductSnapshotDate() {
+		return productSnapshotDate;
+	}
+
+	public String getProductType() {
+		return productType;
+	}
+}
